+++ core/src/com/klimavicius/shooter_game/utils/Constants.java
@@ -3,6 +3,6 @@ package com.klimavicius.shooter_game.utils;
 public class Constants {
     public static final int SCREEN_WIDTH = 1600;
     public static final int SCREEN_HEIGHT = 900;
-    public static final int PLAYER_SPEED = 500;
-    public static final int CAMERA_SPEED = 300;
+    public static final float PLAYER_SPEED = 500;
+    public static final float CAMERA_SPEED = 300;
 }
